--- widgets/client/booking/components/common/index.ts
@@ -1,5 +1,0 @@
-import BackButton from './Button';
-import Navigation from './Header';
-import FilterableList from './FilterableList';
-
-export { BackButton, Navigation, FilterableList };
