--- widgets/client/booking/components/index.ts
@@ -1,6 +1,0 @@
-import App from './App';
-import Intro from './Intro';
-import Booking from './Booking';
-import Product from './Product';
-
-export { App, Intro, Booking, Product };
