--- widgets/client/booking/containers/index.ts
@@ -1,7 +1,0 @@
-import App from './App';
-import Intro from './Intro';
-import Booking from './Booking';
-import CategoryDetail from './CategoryDetail';
-import Product from './Product';
-
-export { App, Intro, Booking, CategoryDetail, Product };
